Remove duplicate staticmethods, move functions to ``cast`` module.
